pyproject.toml
[project]
name = "pydantic-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.29.5",
    "lancedb>=0.24.0",
    "llama-cloud-services>=0.6.41",
    "llama-index>=0.12.46",
    "ollama>=0.5.1",
    "pydantic-ai>=0.2.4",
    "python-dotenv>=1.1.0",
]
